+++ library/src/test/java/com/getstream/sdk/chat/utils/livedata/TestObserver.kt
@@ -1,0 +1,16 @@
+package com.getstream.sdk.chat.utils.livedata
+
+import androidx.lifecycle.Observer
+
+class TestObserver<T> : Observer<T> {
+    var lastObservedValue: T? = null
+        private set
+
+    override fun onChanged(value: T?) {
+        lastObservedValue = value
+    }
+
+    fun reset() {
+        lastObservedValue = null
+    }
+}
